Revert "chore: 프로젝트 요약 컴포넌트의 variant 타입 위치 변경 (#52)"
This reverts commit 4185b515bf181597fa5a193a76183d6ccd08082f.
